Forgot Password Functionality (sort-of) & Registration Page Changes
Added functionality to the forgot password button. The actual forgot password stuff doesn't work (out of the project's scope due to time), but it will allow you to put in an email and "send" instructions on how to change the password. Additionally, I added a bit more checks to make sure information entered in the registration page is acceptable.