Doctor's Patient List & Appointment Patch
Implemented the ability for doctors to view their list of patients, and add to that list. Also implemented a way to handle multiple patients with the same first/last name.

This commit includes a patch to appointment creation that does not allow them to be made for past dates/times, nor appointments on the same day.